weber/requirements.txt
flask
requests
python-jenkins
